requirements.txt
python-docx>=0.8.11
pyinstaller>=5.0.0
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
PyMySQL>=1.1.0
APScheduler>=3.10.0

